requirements.txt
PyYAML
discord
asyncpg
py-dateutil
pycountry
nsfw_dl
